rename const.ts => settings.ts
