+++ project/src/settings.ts
@@ -1,4 +1,4 @@
-
+import { OfferType } from './types/offer-type';
 
 export enum AuthorizationStatus {
   Yes = 'YES',
@@ -40,3 +40,19 @@ export const DEFAULT_MAP_SETTINGS = {
 };
 
 
+export const SortsName = {
+  POPULAR: 'Popular',
+  PRICE_LOW_TO_HIGH: 'Price: low to high',
+  PRICE_HIGH_TO_LOW: 'Price: high to low',
+  TOP_RATED_FIRST: 'Top rated first',
+};
+
+export const SortsRules = {
+  [SortsName.POPULAR]: (a: OfferType, b: OfferType) => a.id - b.id,
+  [SortsName.PRICE_LOW_TO_HIGH]: (a: OfferType, b: OfferType) =>
+    a.price - b.price,
+  [SortsName.PRICE_HIGH_TO_LOW]: (a: OfferType, b: OfferType) =>
+    b.price - a.price,
+  [SortsName.TOP_RATED_FIRST]: (a: OfferType, b: OfferType) =>
+    b.rating - a.rating,
+};
